ACI-Automate-Build-Test/atag/requirements.txt
openpyxl==2.4.8
jinja2
pyyaml
future
requests